requirements.txt
requests>=2.28.0
websocket-client>=1.4.0
urllib3>=1.26.0
schedule>=1.2.0 